[BEAM-4443] Javadods updates. Several classes moved to more fitting packages.
