+++ sdks/java/extensions/euphoria/euphoria-core/src/main/java/org/apache/beam/sdk/extensions/euphoria/core/client/operator/FullJoin.java
@@ -27,6 +27,7 @@ import org.apache.beam.sdk.extensions.euphoria.core.client.functional.BinaryFunc
 import org.apache.beam.sdk.extensions.euphoria.core.client.functional.UnaryFunction;
 import org.apache.beam.sdk.extensions.euphoria.core.client.operator.Join.BuilderParams;
 import org.apache.beam.sdk.extensions.euphoria.core.client.operator.Join.Type;
+import org.apache.beam.sdk.transforms.windowing.WindowFn;
 
 /**
  * Full outer join of two input datasets producing single new dataset.
@@ -44,11 +45,12 @@ import org.apache.beam.sdk.extensions.euphoria.core.client.operator.Join.Type;
  * elements into keys
  * <li>{@code using ....................} {@link BinaryFunctor} receiving left and right element
  * from joined window
- * <li>{@code [windowBy] ...............} windowing function (see {@link Windowing}), default
- * attached windowing
+ * <li>{@code [windowBy] ...............} windowing (see {@link WindowFn}), default is no windowing
+ * <li>{@code [triggeredBy] ............} defines windowing trigger, follows [windowBy] if called
+ * <li>{@code [accumulationMode] .......} windowing accumulation mode, follows [triggeredBy]
  * <li>{@code (output | outputValues) ..} build output dataset
  * </ol>
- */ //TODO update javadoc
+ */
 @Audience(Audience.Type.CLIENT)
 public class FullJoin {
