+++ sdks/java/extensions/euphoria/euphoria-core/src/main/java/org/apache/beam/sdk/extensions/euphoria/core/client/operator/WindowingParams.java
@@ -19,17 +19,20 @@ package org.apache.beam.sdk.extensions.euphoria.core.client.operator;
 
 import javax.annotation.Nullable;
 import org.apache.beam.sdk.extensions.euphoria.core.client.dataset.windowing.Windowing;
+import org.apache.beam.sdk.extensions.euphoria.core.client.operator.base.WindowAware;
 import org.apache.beam.sdk.extensions.euphoria.core.client.operator.windowing.WindowingDesc;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
 import org.apache.beam.sdk.transforms.windowing.Trigger;
 import org.apache.beam.sdk.transforms.windowing.WindowFn;
 import org.apache.beam.sdk.values.WindowingStrategy;
 
+/**
+ * Base for those operator builders which needs to support windowing.
+ */
 class WindowingParams<W extends BoundedWindow>
-implements WindowAware<Object, W>{
+    implements WindowAware<Object, W> {
 
   WindowFn<Object, W> windowFn;
-  //TODO do we always need to set following fields explicitly ? No default values ?
   Trigger trigger;
   WindowingStrategy.AccumulationMode accumulationMode;
